Agent-operating-the-computer-while-on-a-phone-call/computer_tool/pyproject.toml
[project]
name = "computer-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "browser-use>=0.5.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
